Begin rerouting...
begun rerouting of things inline with new plan for compiler to be in
charge of all compiling incl
the loading of plugin, etc... index.html shows an example of use in
stylesheet:

@-plugins-require pluginurl;

we now use a vendor prefix (-plugins-*) for all of our stuff...

interpreted mode just requires jquery 1.5+... simple as that for now.

interpreted adapter is shrunk to almost nothing - not using storage
anymore for now.

interpreted is kicked off by the simple act of including the
-plugins-interpret  attribute on your link/style tag.