Add some necessary meta-data for bug fix, make more tolerant of bad selectors, and additional emulation support
Recent reworks of compiler, etc caused breakage for certain kinds of
automatically updating features... For example, those that would have
been dependent on considering the base in a browser which does not
support an -any- natively.  This adds meta-data to segIndex which
allows
it to be considered again.

Also, the initial search of the tree was made tolerant of a bad selector
the other day, but dynamic was not - this corrects that oversight.

Finally, dynamic attribute changes will now work in IE8 if you are
using
jQuery 1.7