Add queue/common ancestor checking
It is actually possible for things to change while we are processing
the tree... We can't lose those changes.  HTML hitches purposely modify
the tree and while in theory filters shouldn't change the tree, even
potentially keeping track of why the tree changed would be a nightmare.

True, this is a footgun for author of hitches, but not so much for
users of hitches, so - fine.