Add some to the scriptTag strategy to deal with some cases in IE10
IE10 adds some funkiness with script tags…. It had onload support but
was released kinda buggy - likewise, readystates behavior varries
slightly from 9.  The only thing we can't seem to get this to work on
is Windows Phone 7 with IE9 :\ but we've never had proof that it worked
there in the first place so this isn't a regression.