Make it work at least temporarily in Moz
Moz lacks mutation record generation during parse, it seems to be the
only modern browser.  While this has been an open issue for some time,
they are unclear on whether they want to support it.  For now, just
falling back with a minimal ‘wait until domcontentloaded for moz case’…
Not really sure how to handle this better since this is really an
implementation difference in the reading of the spec or desire to
support - I see no way for us to know other than perhaps to set a
timeout and see if we receive some mutations… For now, this will work.

Also corrected crappy 2 space indentation which drives me insane.